Implement hard strategy
Create abstract interface for ai_strategies.
Move existing strategy to “hard” strategy.
Move & update tests
Update existing ai player to use hard strategy by default.
